Switch to C99 data types. Game now works on 64-bits platforms.

+++ src/extern.h
@@ -1,8 +1,8 @@
 extern SDL_Surface *g_screen;
 extern int g_scene;
 
-extern long gameflag[1024];
-extern long gameflag2[1024];
+extern int32_t gameflag[1024];
+extern int32_t gameflag2[1024];
