+++ src/ram.c
@@ -5,9 +5,9 @@
 #include "define.h"
 
 SDL_Surface *g_screen;
-int g_scene;
-long gameflag[1024];
-long gameflag2[1024];
+int32_t g_scene;
+int32_t gameflag[1024];
+int32_t gameflag2[1024];
